Move the mux extensions onto the root channel.
Before, the extensions where mounted on the stream and not on a channel.
Now, the extensions are mounted on the channel that is opened with the
root key. This makes it possible to mount the mux extensions onto any
existing hypercore-protocol stream, multifeed does not have to own the
stream. This will allow to make multifeed compatible with corestore and
corestore-swarm-networker.