fix: Add storage index for persistent tracking of feeds w/ feed removal
With the addition of `removeFeed()`, the previous paradigm of always
numbering storage directories starting at `0` and incrementing with each
feed no longer works as this would have stopped when it encountered a
deleted storage's number. Instead now all storage directories are stored
in an index file called `dirs` in a storage dir called `index`. The
index is updated whenever a feed is added or removed.

This index is agnostic to the name of the dirs except for the assumption
that there is no commas in the name as the dir names are stored with
commas delimiting the names.

Directory names are still incremented numbers. To support the fact
that they can be removed and so the number of feeds tracked does not
match the biggest number used for the directory names, the max number
used is tracked and all new feeds are given a number based on the max.

In the encryption key storage test in the regression tests, the number
of resources was incremented to account for the addition of the index
storage.