fix: Clear storage when writing w/ `writeStringStorage()`
As the previous comment noted, writing `writeStringStorage()` didn't
account for writing to a storage with existing data that was longer than
the current data. This wasn't a problem until the storage index was
implemented. If feeds are removed, then it is very likely that the
current data is shorter than previous data.

To fix this the storage size is retrieved. If it errors while `stat`ing
the file, then its assumed to be non-existent and we just go ahead and
write to the file.